refactor!: move `metrics_manager` to `local\metrics_manager`
Later commmit will introduce `registered_metrics` consumable interface without mutation/management methods.
